Create SiteScan task to be shared between BE components
